Proyecto LLM/requirements.txt
Flask
Flask-Cors
requests
psycopg2-binary
python-dotenv
pytest